==> notebooks/07_Single_Predictor_Logit_Models.ipynb <==
# Logistic Regression Analysis

### Import data and libraries

## Load modules
import pandas as pd
import numpy as np
import seaborn as sns
import statsmodels.api as sm
import statsmodels.formula.api as smf
import os
from matplotlib import pyplot as plt

## Specify WD

wd = (
    "/Users/adam/Documents/BINF667013_BigDataAnalyticsHealthcare/" +
    "Final_Project/TEDS_Study"
)
wd = "/home/abf/BINF667013_Final_Project/"
os.chdir(wd)

## Load data
# teds_laws = pd.read_csv("analysis_objects/teds_laws.csv")
teds_laws = pd.read_csv("analysis_objects/teds_laws.csv")



### Logit model to estimate influence of IC_laws on relapse rate

## Index casese of alcohol use at intake
alc_cases = teds_laws.alc_cases == 1

## Fit Model
alc_result = smf.logit("alc_relapse ~ IC_law", data=teds_laws.loc[alc_cases, :]).fit()
print(alc_result.summary())



## Index casese of heroin use at intake
hrn_cases = teds_laws.hrn_cases == 1

## Fit Model
hrn_result = smf.logit("hrn_relapse ~ IC_law", data=teds_laws.loc[hrn_cases, :]).fit()
print(hrn_result.summary())



## Index casese of methamphetamine use at intake
met_cases = teds_laws.met_cases == 1

## Fit Model
met_result = smf.logit("met_relapse ~ IC_law", data=teds_laws.loc[met_cases, :]).fit()
print(met_result.summary())



### Logit model to estimate influence of Unborn Child Protective Commitment on PREG Admissions

## Index casese of alcohol use at intake
fem_alc_cases = teds_laws[
    (teds_laws['alc_cases'] == 1) & (~pd.isna(teds_laws.PREG))
].index
teds_laws.loc[fem_alc_cases,:].shape

## Fit Model
alc_result = smf.logit(
    "PREG ~ Protect_Unborn", 
    data=teds_laws.loc[fem_alc_cases, :].rename(
        {"IC_Circumstances_Protect an unborn child":"Protect_Unborn"}, axis=1
    ).assign(PREG = lambda x: x['PREG'] -1)
).fit()
print(alc_result.summary())

## Index casese of alcohol use at intake
fem_hrn_cases = teds_laws[
    (teds_laws['hrn_cases'] == 1) & (~pd.isna(teds_laws.PREG))
].index
teds_laws.loc[fem_alc_cases,:].shape

## Fit Model
hrn_result = smf.logit(
    "PREG ~ Protect_Unborn", 
    data=teds_laws.loc[fem_hrn_cases, :].rename(
        {"IC_Circumstances_Protect an unborn child":"Protect_Unborn"}, axis=1
    ).assign(PREG = lambda x: x['PREG'] -1)
).fit()
print(hrn_result.summary())

## Index casese of alcohol use at intake
fem_met_cases = teds_laws[
    (teds_laws['met_cases'] == 1) & (~pd.isna(teds_laws.PREG))
].index
teds_laws.loc[fem_alc_cases,:].shape

## Fit Model
met_result = smf.logit(
    "PREG ~ Protect_Unborn", 
    data=teds_laws.loc[fem_met_cases, :].rename(
        {"IC_Circumstances_Protect an unborn child":"Protect_Unborn"}, axis=1
    ).assign(PREG = lambda x: x['PREG'] -1)
).fit()
print(met_result.summary())

### Logit model to estimate the influence of IC_Law Status on Dropout vs Completion 

### Select Cases
dropout_alc_cases = teds_laws[
    (teds_laws['alc_cases'] == 1) & (teds_laws.REASON.isin([1,2]))
].index

dropout_hrn_cases = teds_laws[
    (teds_laws['hrn_cases'] == 1) & (teds_laws.REASON.isin([1,2]))
].index

dropout_met_cases = teds_laws[
    (teds_laws['met_cases'] == 1) & (teds_laws.REASON.isin([1,2]))
].index

teds_laws.loc[dropout_alc_cases, :].shape

## Fit Models
alc_result = smf.mnlogit("REASON ~ IC_law", data=teds_laws.loc[dropout_alc_cases, :]).fit()
print(alc_result.summary())

## Fit Model
hrn_result = smf.mnlogit("REASON ~ IC_law", data=teds_laws.loc[dropout_hrn_cases, :]).fit()
print(hrn_result.summary())

## Fit Model
met_result = smf.mnlogit("REASON ~ IC_law", data=teds_laws.loc[dropout_met_cases, :]).fit()
print(met_result.summary())

### Multinomial Logistic Regression to estimate IC Law influence on referral source


## Fit Model
alc_result = smf.mnlogit("PSOURCE ~ IC_law", data=teds_laws.loc[alc_cases, :]).fit()
print(alc_result.summary())

## Fit Model
hrn_result = smf.mnlogit("PSOURCE ~ IC_law", data=teds_laws.loc[hrn_cases, :]).fit()
print(hrn_result.summary())

## Fit Model
met_result = smf.mnlogit("PSOURCE ~ IC_law", data=teds_laws.loc[met_cases, :]).fit()
print(met_result.summary())




### Multinomial Logistic Regression to estimate Influence of IC Laws on Services 

## Verify that the data is well defined,
## all cases represented by at least one observation
## to avoid division by zero
teds_laws.loc[alc_cases, ["IC_law","SERVICES"]].dropna().value_counts().sort_index()

## Fit Model -- for some stupid reason, the formula syntax fails here
alc_result = sm.MNLogit(endog=teds_laws.loc[alc_cases,["SERVICES"]], exog=teds_laws.loc[alc_cases,["IC_law"]]).fit()
print(alc_result.summary())

### Note that the direction appears to be flipped in this analysis.  

## Fit Model
hrn_result = smf.mnlogit("SERVICES ~ IC_law", data=teds_laws.loc[hrn_cases, :]).fit()
print(hrn_result.summary())

## Fit Model
met_result = smf.mnlogit("SERVICES ~ IC_law", data=teds_laws.loc[met_cases, :]).fit()
print(met_result.summary())

### Multinomial Logistic Regression to estimate Influence of IC Laws on Reasons for Discharge 


## Fit Model
alc_result = smf.mnlogit("REASON ~ IC_law", data=teds_laws.loc[alc_cases, :]).fit()
print(alc_result.summary())

## Fit Model
hrn_result = smf.mnlogit("REASON ~ IC_law", data=teds_laws.loc[hrn_cases, :]).fit()
print(hrn_result.summary())

## Fit Model
met_result = smf.mnlogit("REASON ~ IC_law", data=teds_laws.loc[met_cases, :]).fit()
print(met_result.summary())
